Safari-611.1.10/LayoutTests/inspector/dom/attributeModified-expected.txt

Tests for the DOM.attributeModified event.


== Running test suite: DOM.attributeModified
-- Running test case: DOM.attributeModified.WithNamespace
Creating test node...
{
  "id": "with-namespace",
  "attribute:existing": "42"
}

Adding attribute with namespace...
{
  "id": "with-namespace",
  "attribute:existing": "42",
  "attribute:new": "1"
}

Replacing attribute with namespace...
{
  "id": "with-namespace",
  "attribute:existing": "42",
  "attribute:new": "2"
}

Replacing attribute without namespace...
{
  "id": "with-namespace",
  "attribute:existing": "42",
  "attribute:new": "3"
}
